Move `my-account` to `account`
