+++ src/pages/account.tsx
@@ -1,42 +1,68 @@
-import {
-  ContentTypes,
-  getTranslations,
-  useContentful,
-} from '@/utils/hooks/useContentful'
-import { GetStaticPropsContext } from 'next'
-import { useAccount } from 'wagmi'
-import { useGetNfts } from '@/utils/hooks/useGetNfts'
-import { NftCard } from '@/components/NftCard'
-import { useGetNftsCollections } from '@/utils/hooks/useGetNftsCollection'
-import { ourCollectionsAddresses } from '@/consts'
-import { compareNfts } from '@/utils/helpers'
+import RewardCard from '@/components/account/RewardCard'
+import { ContentTypes, getTranslations } from '@/utils/hooks/useContentful'
+import { Stack } from '@mui/material'
+import { GetStaticProps, GetStaticPropsContext } from 'next'
+import React from 'react'
 
-const Collections = () => {
-  const translate = useContentful(ContentTypes.collectionsPage)
-  const { address } = useAccount()
-  const ownedNfts = useGetNfts(address)
-  const wlfSocietyNfts = useGetNftsCollections(ourCollectionsAddresses)
+const mockCardData = [
+  {
+    title: 'Your contribution',
+    amount: '2.5',
+    isEthAmount: true,
+    buttonText: 'Distribute (soon)',
+    buttonAction: () => {},
+    isDisabled: true,
+  },
+  {
+    title: 'Your rewards',
+    amount: '0.02456',
+    isEthAmount: true,
+    buttonText: 'Claim Rewards',
+    buttonAction: () => {},
+    isDisabled: false,
+  },
+  {
+    title: 'Game Tokens',
+    amount: '250',
+    isEthAmount: false,
+    buttonText: 'Play Game (soon)',
+    buttonAction: () => {},
+    isDisabled: true,
+  },
+]
 
+export const Account = () => {
   return (
-    <>
-      {/* {wlfSocietyNfts?.map((nft, index) => (
-        <NftCard
-          key={index}
-          nft={nft}
-          owned={ownedNfts.some((ownedNft) => compareNfts(ownedNft, nft))}
-        />
-      ))} */}
-    </>
+    <Stack
+      pt={15}
+      pb={10}
+      sx={{ alignItems: 'center', backgroundColor: 'neutral.400' }}
+    >
+      <Stack
+        direction={'row'}
+        sx={{ justifyContent: 'center', background: 'inherit' }}
+        width={'100%'}
+        px={10}
+      >
+        {mockCardData.map((cardData, idx) => {
+          return <RewardCard key={`reward-card-${idx}`} {...cardData} />
+        })}
+      </Stack>
+    </Stack>
   )
 }
 
-export async function getStaticProps({ locale }: GetStaticPropsContext) {
+export const getStaticProps: GetStaticProps<{}> = async ({
+  locale,
+}: GetStaticPropsContext) => {
   return {
     // Will be passed to the page component as props
     props: {
-      translations: await getTranslations(ContentTypes.collectionsPage, locale),
+      translations: await getTranslations(ContentTypes.landingPage, locale),
+      locale,
     },
+    revalidate: 60, // In seconds
   }
 }
 
-export default Collections
+export default Account
